examples/lab/app-producer/pyproject.toml
[project]
name = "plywatch-lab-producer"
version = "0.1.0"
description = "Lab producer app for Plywatch"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "pydantic>=2.0,<3.0",
    "uvicorn>=0.30,<1.0",
    "loom-kernel[celery,config,sqlalchemy]==0.2.1.dev17",
]

[project.scripts]
plywatch-lab-traffic-loop = "app.traffic_loop:main"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
